+++ docs/기능 목록.md
@@ -40,8 +40,9 @@
 
 #### loading / error
 
-- [] skeleton
-- [] suspense
+- [x] skeleton
+
+#### api hook 수정
 
 ---
 
@@ -63,6 +64,4 @@
 ### 수정 사항
 
 - useResponsive / 서버는 페이지의 width를 알 수 있는 방법이 없음
-- index.tsx:110 Error: Loading initial props cancelled
-- eslint 속도
 - OrderAtUI / OrderAtAPI
